+++ epan/dissectors/packet-msrp.c
@@ -483,7 +483,7 @@ dissect_msrp(tvbuff_t *tvb, packet_info *pinfo, proto_tree *tree)
 	gint hf_index;
 	gint value_offset;
 	guchar c;
-	size_t value_len;
+	gint value_len;
 	char *value;
 	gboolean have_body = FALSE;
 	gboolean found_match = FALSE;
